requirements.txt
# ── Cryptography ─────────────────────────────────────────────────────────────
# PyNaCl: Python bindings for libsodium (X25519, Ed25519, XSalsa20-Poly1305)
PyNaCl>=1.5.0
# cryptography: Argon2 KDF + AES-GCM for keystore encryption
cryptography>=42.0.0

# ── Web framework ─────────────────────────────────────────────────────────────
# FastAPI: async REST + WebSocket API server
fastapi>=0.111.0
# uvicorn: ASGI server that runs FastAPI
uvicorn[standard]>=0.29.0

# ── Serialisation ─────────────────────────────────────────────────────────────
# protobuf: Session message payload encoding (used in Layer 4)
protobuf>=4.25.0

# ── HTTP client ───────────────────────────────────────────────────────────────
# httpx: async HTTP client used by the network layer (Layer 3)
httpx>=0.27.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
